Fix Widget rendering on DPL pages
DPL is cloning Parser or using new instance, and therefore ExtensionData is lost and not replaced during final `onParserAfterTidy` on the main run. Is highly unlikely that the module will have same Random string generated to leak to different parser/part of pages, also even now the collision of markers might appear on single page run.